refactor: HomeViewCollectionLayout 로직 분리
